requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
